refactor(libpatchwork): complete libpatchwork to handle new status system

+++ src/libstd/user/common/exit_stack.h
@@ -6,6 +6,6 @@
 
 void _exit_stack_init(void);
 
-uint64_t _exit_stack_push(void (*func)(void));
+int _exit_stack_push(void (*func)(void));
 
 void _exit_stack_dispatch(void);
